requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
git+https://github.com/wazo-platform/wazo-auth-client.git
